cockroachkvs: move from internal/crdbtest
Move the internal/crdbtest package to pebble/cockroachkvs. We plan to begin to
use these Comparer and KeySchemas defined within the Pebble repository from
Cockroach, removing the toil of keeping the two implementations synchronized.
