+++ cmd/pebble/scan.go
@@ -14,7 +14,7 @@ import (
 	"time"
 
 	"github.com/cockroachdb/pebble"
-	"github.com/cockroachdb/pebble/internal/crdbtest"
+	"github.com/cockroachdb/pebble/cockroachkvs"
 	"github.com/cockroachdb/pebble/internal/randvar"
 	"github.com/spf13/cobra"
 )
@@ -73,7 +73,7 @@ func runScan(cmd *cobra.Command, args []string) {
 				b := d.NewBatch()
 				var value []byte
 				for end := i + batch; i < end; i++ {
-					keys[i] = crdbtest.EncodeMVCCKey(nil, encodeUint32Ascending([]byte("key-"), uint32(i)), uint64(i+1), 0)
+					keys[i] = cockroachkvs.EncodeMVCCKey(nil, encodeUint32Ascending([]byte("key-"), uint32(i)), uint64(i+1), 0)
 					value = scanConfig.values.Bytes(rng, value)
 					if err := b.Set(keys[i], value, nil); err != nil {
 						log.Fatal(err)
