+++ cockroachkvs/key_schema_test.go
@@ -1,4 +1,8 @@
-package crdbtest
+// Copyright 2024 The LevelDB-Go and Pebble Authors. All rights reserved. Use
+// of this source code is governed by a BSD-style license that can be found in
+// the LICENSE file.
+
+package cockroachkvs
 
 import (
 	"bytes"
