+++ cockroachkvs/testdata/seek
