+++ cockroachkvs/testdata/suffix_types
